requirements.txt
aiogram==3.13.1
anthropic==0.39.0
aiomysql==0.2.0
python-dotenv==1.0.1
